Add files via upload
This file is related to a memory game, in this game, after running the code, in the first step, a circle turns on and then off, and we have to memorize it, and then specify it with the mouse, and to go to the next step. Press the Enter key.